Traipse Dev 'OpenRPG' {090813-01}
Traipse is a distribution of OpenRPG that is designed to be easy to setup and go. Traipse also makes it easy for developers to work on code without fear of sacrifice. 'Ornery-Orc' continues the trend of 'Grumpy' and adds fixes to the code. 'Ornery-Orc''s main goal is to offer more advanced features and enhance the productivity of the user.

Update Summary:
Minor fix to Update Manager.  Files only shows files added to the most recent commit. Please Manifest your important files when switching branches.  Nodehandler now adds .xml to saved nodes if .xml is not the last 4 characters.