Traipse Alpha 'OpenRPG' {090909-00}
Traipse is a distribution of OpenRPG that is designed to be easy to setup and go. Traipse also makes it easy for developers to work on code without fear of sacrifice. 'Ornery-Orc' continues the trend of 'Grumpy' and adds fixes to the code. 'Ornery-Orc''s main goal is to offer more advanced features and enhance the productivity of the user.

Update Summary:
Fixes problems with Text nodes. Fixes log problem in Fog.  Fixes Mini Lib loading problem.  Fixes problem with whispers in Alias Lib.
Creates new Alpha Branch.